+++ modules/ambient_mode/views/Items.qml
@@ -69,6 +69,19 @@ FocusScope {
         title: moduleRoot.moduleName
     }
 
+    Text {
+        text: "Media Directory: " + ambientModeBackend.mediaRoot()
+        color: root.tertiaryColor
+        font.family: root.globalFont
+        anchors.top: parent.top
+        anchors.left: parent.left
+        anchors.topMargin: root.sh * 0.1958333
+        anchors.leftMargin: root.sw * 0.125
+        width: root.sw * 0.76875
+        elide: Text.ElideRight
+        font.pixelSize: root.sh * 0.0291667
+    }
+
     // Empty state
     Column {
         anchors.centerIn: parent
@@ -84,7 +97,7 @@ FocusScope {
             font.pixelSize: root.sh * 0.05 //24
         }
         Text {
-            text: "Please add items in the ambient:mode media directory"
+            text: "Please add items in the loop media directory"
             color: root.tertiaryColor
             font.family: root.globalFont
             font.capitalization: Font.AllUppercase